requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
werkzeug==3.0.1
email-validator==2.1.0
sib-api-v3-sdk==7.6.0
vonage==3.14.0
plaid-python==20.0.0
gunicorn==21.2.0
